move StaticMeshIntegrator
